feat(preview): add preview for new homepage variant (#1485)
* feat(herobody): update variant in frontmatter

* refactor(herosection): shift to ts

* refactor(herosection): shift stuff out

* refactor(hero): add css and update import

* fix(hero): style fixes

* fix(heorbody): update colours

* refactor(herobody): use icon

* refactor(hero): remove gray overlay on image only layout

* fix(homepagepreview): add default if no variant

* fix(use): shifted to previewpanel

* fix(herodropdown): add missinng default

---------
