+++ src/assets/icons/index.ts
@@ -17,3 +17,4 @@ export * from "./BxLifeBuoy"
 export * from "./BxCopy"
 export * from "./BxDraggable"
 export * from "./BxDraggableVertical"
+export * from "./BxGrayTranslucent"
