+++ src/constants/homepage.ts
@@ -1,0 +1,14 @@
+export const HERO_LAYOUTS = {
+  CENTERED: {
+    value: "center",
+    label: "Centre-aligned text",
+  },
+  IMAGE_ONLY: {
+    value: "image",
+    label: "Image only",
+  },
+  SIDE_SECTION: {
+    value: "side",
+    label: "Side section",
+  },
+} as const
