+++ src/styles/isomer-template.scss
@@ -1,4 +1,5 @@
 @charset "UTF-8";
+
 .bg-media-color-0 {
   background-color: #4b268e;
 }
@@ -10886,7 +10887,6 @@ a.navbar-link:hover {
 }
 
 .bp-hero-body {
-  background: rgba(0, 0, 0, 0.25);
   flex-grow: 1;
   flex-shrink: 0;
   padding: 3rem 1.5rem;
