+++ src/styles/templates/components/hero.scss
@@ -1,0 +1,7 @@
+.min-height-mobile {
+  min-height: 398px;
+}
+
+.gray-overlay {
+  background: rgba(0, 0, 0, 0.25);
+}
